long

+++ ftpslave_utils.c
@@ -70,7 +70,7 @@ int resend_file(char * name, long size, int clientfd){
 
     Rio_readinitb(&fileAsked, fd);
     while ((n=Rio_readnb(&fileAsked,buf,BUF_LEN)) > 0) {
-      printf(">Server send %ld bytes\n",n);
+      printf(">Server send %ld bytes\n",(long)n);
       Rio_writen(clientfd, buf, n);
     }
     puts(">END");
